Moved openDatabases to JabRefFrame
