+++ src/main/java/org/jabref/gui/JabRefFrame.java
@@ -2,6 +2,8 @@ package org.jabref.gui;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
@@ -41,6 +43,7 @@ import org.jabref.gui.desktop.JabRefDesktop;
 import org.jabref.gui.help.HelpAction;
 import org.jabref.gui.importer.ImportEntriesDialog;
 import org.jabref.gui.importer.NewEntryAction;
+import org.jabref.gui.importer.ParserResultWarningDialog;
 import org.jabref.gui.importer.actions.OpenDatabaseAction;
 import org.jabref.gui.keyboard.KeyBinding;
 import org.jabref.gui.keyboard.KeyBindingRepository;
@@ -58,6 +61,9 @@ import org.jabref.logic.help.HelpFile;
 import org.jabref.logic.importer.ImportCleanup;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.shared.DatabaseNotSupportedException;
+import org.jabref.logic.shared.exception.InvalidDBMSConnectionPropertiesException;
+import org.jabref.logic.shared.exception.NotASharedDatabaseException;
 import org.jabref.logic.undo.AddUndoableActionEvent;
 import org.jabref.logic.undo.UndoChangeEvent;
 import org.jabref.logic.undo.UndoRedoEvent;
@@ -848,6 +854,120 @@ public class JabRefFrame extends BorderPane implements LibraryTabContainer {
         globalSearchBar.updateHintVisibility();
         getLibraryTabs().forEach(LibraryTab::setupMainPanel);
         getLibraryTabs().forEach(tab -> tab.getMainTable().getTableModel().resetFieldFormatter());
+    }
+
+    void openDatabases(List<ParserResult> parserResults, boolean isBlank) {
+        final List<ParserResult> failed = new ArrayList<>();
+        final List<ParserResult> toOpenTab = new ArrayList<>();
+
+        // If the option is enabled, open the last edited libraries, if any.
+        if (!isBlank && prefs.getWorkspacePreferences().shouldOpenLastEdited()) {
+            openLastEditedDatabases();
+        }
+
+        // From here on, the libraries provided by command line arguments are treated
+
+        // Remove invalid databases
+        List<ParserResult> invalidDatabases = parserResults.stream()
+                                                           .filter(ParserResult::isInvalid)
+                                                           .toList();
+        failed.addAll(invalidDatabases);
+        parserResults.removeAll(invalidDatabases);
+
+        // passed file (we take the first one) should be focused
+        Path focusedFile = parserResults.stream()
+                                        .findFirst()
+                                        .flatMap(ParserResult::getPath)
+                                        .orElse(prefs.getGuiPreferences()
+                                                     .getLastFocusedFile())
+                                        .toAbsolutePath();
+
+        // Add all bibDatabases databases to the frame:
+        boolean first = false;
+        for (ParserResult parserResult : parserResults) {
+            // Define focused tab
+            if (parserResult.getPath().filter(path -> path.toAbsolutePath().equals(focusedFile)).isPresent()) {
+                first = true;
+            }
+
+            if (parserResult.getDatabase().isShared()) {
+                try {
+                    OpenDatabaseAction.openSharedDatabase(
+                            parserResult,
+                            this,
+                            getDialogService(),
+                            prefs,
+                            Globals.stateManager,
+                            Globals.entryTypesManager,
+                            fileUpdateMonitor,
+                            getUndoManager(),
+                            Globals.TASK_EXECUTOR);
+                } catch (
+                        SQLException |
+                        DatabaseNotSupportedException |
+                        InvalidDBMSConnectionPropertiesException |
+                        NotASharedDatabaseException e) {
+
+                    LOGGER.error("Connection error", e);
+                    getDialogService().showErrorDialogAndWait(
+                            Localization.lang("Connection error"),
+                            Localization.lang("A local copy will be opened."),
+                            e);
+                    toOpenTab.add(parserResult);
+                }
+            } else if (parserResult.toOpenTab()) {
+                // things to be appended to an opened tab should be done after opening all tabs
+                // add them to the list
+                toOpenTab.add(parserResult);
+            } else {
+                addTab(parserResult, first);
+                first = false;
+            }
+        }
+
+        // finally add things to the currently opened tab
+        for (ParserResult parserResult : toOpenTab) {
+            addTab(parserResult, first);
+            first = false;
+        }
+
+        for (ParserResult pr : failed) {
+            String message = Localization.lang("Error opening file '%0'",
+                    pr.getPath().map(Path::toString).orElse("(File name unknown)")) + "\n" +
+                    pr.getErrorMessage();
+
+            getDialogService().showErrorDialogAndWait(Localization.lang("Error opening file"), message);
+        }
+
+        // Display warnings, if any
+        for (int tabNumber = 0; tabNumber < parserResults.size(); tabNumber++) {
+            // ToDo: Method needs to be rewritten, because the index of the parser result and of the libraryTab may not
+            //  be identical, if there are also other tabs opened, that are not libraryTabs. Currently there are none,
+            //  therefore for now this ok.
+            ParserResult pr = parserResults.get(tabNumber);
+            if (pr.hasWarnings()) {
+                ParserResultWarningDialog.showParserResultWarningDialog(pr, getDialogService());
+                showLibraryTabAt(tabNumber);
+            }
+        }
+
+        // After adding the databases, go through each and see if
+        // any post open actions need to be done. For instance, checking
+        // if we found new entry types that can be imported, or checking
+        // if the database contents should be modified due to new features
+        // in this version of JabRef.
+        parserResults.forEach(pr -> OpenDatabaseAction.performPostOpenActions(pr, getDialogService()));
+
+        LOGGER.debug("Finished adding panels");
+    }
+
+    private void openLastEditedDatabases() {
+        List<Path> lastFiles = prefs.getGuiPreferences().getLastFilesOpened();
+        if (lastFiles.isEmpty()) {
+            return;
+        }
+
+        getOpenDatabaseAction().openFiles(lastFiles);
     }
 
     /**
