+++ src/main/java/org/jabref/gui/JabRefGUI.java
@@ -1,8 +1,5 @@
 package org.jabref.gui;
 
-import java.nio.file.Path;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,15 +11,10 @@ import javafx.stage.Stage;
 
 import org.jabref.gui.help.VersionWorker;
 import org.jabref.gui.icon.IconTheme;
-import org.jabref.gui.importer.ParserResultWarningDialog;
-import org.jabref.gui.importer.actions.OpenDatabaseAction;
 import org.jabref.gui.keyboard.TextInputKeyBindings;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.net.ProxyRegisterer;
-import org.jabref.logic.shared.DatabaseNotSupportedException;
-import org.jabref.logic.shared.exception.InvalidDBMSConnectionPropertiesException;
-import org.jabref.logic.shared.exception.NotASharedDatabaseException;
 import org.jabref.logic.util.WebViewStore;
 import org.jabref.model.strings.StringUtil;
 import org.jabref.model.util.FileUpdateMonitor;
@@ -46,8 +38,6 @@ public class JabRefGUI {
     private final List<ParserResult> parserResults;
     private final boolean isBlank;
     private boolean correctedWindowPos;
-    private final List<ParserResult> failed = new ArrayList<>();
-    private final List<ParserResult> toOpenTab = new ArrayList<>();
 
     public JabRefGUI(Stage mainStage,
                      List<ParserResult> parserResults,
@@ -165,7 +155,7 @@ public class JabRefGUI {
                 event.consume();
             }
         });
-        Platform.runLater(this::openDatabases);
+        Platform.runLater(() -> mainFrame.openDatabases(parserResults, isBlank));
 
         if (!(fileUpdateMonitor.isActive())) {
             getMainFrame().getDialogService()
@@ -173,110 +163,6 @@ public class JabRefGUI {
                                   "and processes and restart. You may encounter errors if you continue " +
                                   "with this session."));
         }
-    }
-
-    /**
-     * ToDo: This should be part of JabRefFrame
-     */
-    private void openDatabases() {
-        // If the option is enabled, open the last edited libraries, if any.
-        if (!isBlank && preferencesService.getWorkspacePreferences().shouldOpenLastEdited()) {
-            openLastEditedDatabases();
-        }
-
-        // From here on, the libraries provided by command line arguments are treated
-
-        // Remove invalid databases
-        List<ParserResult> invalidDatabases = parserResults.stream()
-                                                           .filter(ParserResult::isInvalid)
-                                                           .toList();
-        failed.addAll(invalidDatabases);
-        parserResults.removeAll(invalidDatabases);
-
-        // passed file (we take the first one) should be focused
-        Path focusedFile = parserResults.stream()
-                                        .findFirst()
-                                        .flatMap(ParserResult::getPath)
-                                        .orElse(preferencesService.getGuiPreferences()
-                                                                 .getLastFocusedFile())
-                                        .toAbsolutePath();
-
-        // Add all bibDatabases databases to the frame:
-        boolean first = false;
-        for (ParserResult parserResult : parserResults) {
-            // Define focused tab
-            if (parserResult.getPath().filter(path -> path.toAbsolutePath().equals(focusedFile)).isPresent()) {
-                first = true;
-            }
-
-            if (parserResult.getDatabase().isShared()) {
-                try {
-                    OpenDatabaseAction.openSharedDatabase(
-                            parserResult,
-                            mainFrame,
-                            mainFrame.getDialogService(),
-                            preferencesService,
-                            Globals.stateManager,
-                            Globals.entryTypesManager,
-                            fileUpdateMonitor,
-                            mainFrame.getUndoManager(),
-                            Globals.TASK_EXECUTOR);
-                } catch (SQLException |
-                        DatabaseNotSupportedException |
-                        InvalidDBMSConnectionPropertiesException |
-                        NotASharedDatabaseException e) {
-
-                    LOGGER.error("Connection error", e);
-                    mainFrame.getDialogService().showErrorDialogAndWait(
-                            Localization.lang("Connection error"),
-                            Localization.lang("A local copy will be opened."),
-                            e);
-                    toOpenTab.add(parserResult);
-                }
-            } else if (parserResult.toOpenTab()) {
-                // things to be appended to an opened tab should be done after opening all tabs
-                // add them to the list
-                toOpenTab.add(parserResult);
-            } else {
-                mainFrame.addTab(parserResult, first);
-                first = false;
-            }
-        }
-
-        // finally add things to the currently opened tab
-        for (ParserResult parserResult : toOpenTab) {
-            mainFrame.addTab(parserResult, first);
-            first = false;
-        }
-
-        for (ParserResult pr : failed) {
-            String message = Localization.lang("Error opening file '%0'",
-                    pr.getPath().map(Path::toString).orElse("(File name unknown)")) + "\n" +
-                    pr.getErrorMessage();
-
-            mainFrame.getDialogService().showErrorDialogAndWait(Localization.lang("Error opening file"), message);
-        }
-
-        // Display warnings, if any
-        for (int tabNumber = 0; tabNumber < parserResults.size(); tabNumber++) {
-            // ToDo: Method needs to be rewritten, because the index of the parser result and of the libraryTab may not
-            //  be identical, if there are also other tabs opened, that are not libraryTabs. Currently there are none,
-            //  therefore for now this ok.
-            ParserResult pr = parserResults.get(tabNumber);
-            if (pr.hasWarnings()) {
-                ParserResultWarningDialog.showParserResultWarningDialog(pr, mainFrame.getDialogService());
-                mainFrame.showLibraryTabAt(tabNumber);
-            }
-        }
-
-        // After adding the databases, go through each and see if
-        // any post open actions need to be done. For instance, checking
-        // if we found new entry types that can be imported, or checking
-        // if the database contents should be modified due to new features
-        // in this version of JabRef.
-        parserResults.forEach(pr -> OpenDatabaseAction.performPostOpenActions(pr, mainFrame.getDialogService()));
-
-        LOGGER.debug("Finished adding panels");
     }
 
     private void saveWindowState(Stage mainStage) {
@@ -316,15 +202,6 @@ public class JabRefGUI {
         return !Screen.getPrimary().getBounds().contains(
                 preferencesService.getGuiPreferences().getPositionX(),
                 preferencesService.getGuiPreferences().getPositionY());
-    }
-
-    private void openLastEditedDatabases() {
-        List<Path> lastFiles = preferencesService.getGuiPreferences().getLastFilesOpened();
-        if (lastFiles.isEmpty()) {
-            return;
-        }
-
-        getMainFrame().getOpenDatabaseAction().openFiles(lastFiles);
     }
 
     public static JabRefFrame getMainFrame() {
